fix: production-readiness corrections from automated review
- searchAll: per-type appliedFilters (no false claims on artist results);
  stop sending tag/year filters /api/artist silently ignores
- listRecentlyPlayed/listMostPlayed/listTopRated: fetch _start=0 and apply
  offset after client-side filtering so pagination no longer drops rows
- listTopRated: add hasMore/partial signal for capped/under-delivered windows
- assorted high/critical fixes across tools, services, schemas, and client

+++ src/schemas/common.ts
@@ -123,7 +123,7 @@ export const EnhancedSearchSchema = SearchQuerySchema.extend({
 });
 
 // Rating validation
-export const RatingSchema = z.number().min(0).max(5);
+export const RatingSchema = z.number().int().min(0).max(5);
 
 // Duration validation for timeouts
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type,@typescript-eslint/explicit-module-boundary-types -- schema factory; return type inferred by zod, explicit annotation would be unwieldy
